requirements.txt
speedtest-cli
ping3
